fix(source): stop is_alloc_limit depending on a feature flag, and stop it overclaiming (#686)
Five from the systems and safety lens of the panel review, all of which it
demonstrated by running rather than by reading.

The one that matters most: is_alloc_limit gave a different answer for the
same value depending on whether the jxl feature was on. JxlError and
DecoderAllocLimitExceeded are declared unconditionally, and #634's doc
promises a caller's match has the same arms in either build, but I put a
cfg on the predicate's arm. Reproduced:

  jxl feature = false, is_alloc_limit = false
  jxl feature = true,  is_alloc_limit = true

on the identical value. Features are additive, so one crate in a workspace
turning jxl on would silently change another crate's error handling. The
cfg is gone and it costs nothing, because the variant was always there.

That got in because the arm had no test. Deleting the whole arm left the
suite green with the feature on, since the jxl row in the refusal table is
16x16 on purpose so it trips the declared-geometry check and never reaches
the tracker. The new test builds the value by hand rather than decoding,
which is also the only way to assert it in the build that has no decoder.

The doc and the CHANGELOG said "every shape the budget can refuse in" and
there is a fourth. Raster::ppm_load, csv_load and matrix_load are public
decode entry points returning this same enum, and an over-large declared
geometry comes back as Raster(ByteBudgetExceeded) whose message reads
"needs 10800000000 bytes, exceeding the 8589934592-byte allocation budget".
It looks exactly like this and it is not: that budget is
DEFAULT_MAX_ALLOC_BYTES, the raster construction ceiling, not
DecodeLimits::max_alloc_bytes, so raising the decode limit does nothing
about it. The predicate's whole test is "does raising max_alloc_bytes fix
this", all three exclusions fail it, and there is now a control asserting
the message reads like an allocation refusal before asserting the predicate
says no.

check_image_alloc takes u64 band and sample counts now. Every saturation in
this area goes up, to u64::MAX, which exceeds_alloc_budget treats as a
sentinel refusal; the u32 parameters made EXR's channel count saturate the
price down instead, which is the one direction that turns a refusal into a
decode. Only the geometry the message reports narrows now, inside the
shared helper, where a count that large understates the message and cannot
change the verdict.

And the error that reports an allocation refusal no longer allocates to
format itself. The geometry arm went through format! into a String, which
puts a heap allocation on the path formatting an out-of-memory error, where
the host is least likely to serve one. A small Display newtype writes
straight into the formatter.

Three mutations, all caught:

  re-gate the jxl arm behind the feature -> 1 failed
  delete the jxl arm entirely            -> 1 failed in BOTH builds
  sweep the construction budget in       -> 1 failed

Suite 1872 without features, 1898 with --features jxl. Gate clean on all
three feature sets, doc 47.

+++ src/encode_tiff.rs
@@ -647,8 +647,8 @@ fn decode_current_image<R: Read + Seek>(
         "TIFF page pixel buffer",
         width,
         height,
-        channels as u32,
-        u32::from(bit_depth).div_ceil(8),
+        channels as u64,
+        u64::from(bit_depth).div_ceil(8),
     )?;
     let orientation = read_tiff_orientation(decoder);
     let result = decoder.read_image().map_err(tiff_decode_err)?;
